Move distribution classes around.
